requirements.txt
# dgl==0.7.1
pandas==1.1.5
pydot==1.4.2
tensorboard==2.6.0
tensorboardX==2.1
# torch==1.8.0
torchdiffeq==0.2.2
torchviz==0.0.2
matplotlib==3.4.3
requests==2.26.0 
scikit-learn==1.0.1 
statsmodels==0.13.2